requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning libraries
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep learning libraries
tensorflow>=2.8.0
torch>=1.12.0
torchvision>=0.13.0

# Time series analysis
statsmodels>=0.13.0
arch>=5.3.0
pykalman>=0.9.5

# Financial analysis
yfinance>=0.1.70
ta>=0.10.0
pandas-ta>=0.3.14b

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0

# Data processing
openpyxl>=3.0.0
xlrd>=2.0.0

# API and web scraping
requests>=2.27.0
beautifulsoup4>=4.10.0

# Optimization
optuna>=2.10.0
hyperopt>=0.2.7

# Feature engineering
feature-engine>=1.4.0

# Model interpretation
shap>=0.40.0
lime>=0.2.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0 